update print list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	size_t count = 0;
+	unsigned int count = 0;
 	
 	while (h)
 	{
@@ -18,7 +18,7 @@ size_t print_list(const list_t *h)
 		}
 		else
 		{
-			printf("[%u] %s\n", h->len, h->str);
+			printf("[%lu] %s\n", h->len, h->str);
 		}
 		h = h->next;
 		count++;
